task_manager-main/requirements.txt
Flask==3.0.2
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
python-dotenv==1.0.1
Werkzeug==3.0.1
Flask-Limiter==2.8.1